+++ lib/src/segment_default_options.dart
