+++ lib/src/segment_method_channel.dart
@@ -1,7 +1,7 @@
 import 'package:meta/meta.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_segment/segment_default_options.dart';
-import 'package:flutter_segment/segment_platform_interface.dart';
+import 'package:flutter_segment/src/segment_default_options.dart';
+import 'package:flutter_segment/src/segment_platform_interface.dart';
 
 const MethodChannel _channel = MethodChannel('flutter_segment');
